refactor(libflux): move the rust code into a single crate
The rust code is now a single crate named `flux` and it will produce a
`libflux` file for use when we integrate it with Go.

A header file has been added to be used as the C interface that Go will
use. No other modifications, other than a few changes to make it work by
referencing internal crates, have been made to the source code.
