--- internal/rust/parser/src/parser.h
@@ -1,1 +1,0 @@
-void go_parse(const char* s);
