--- internal/rust/scanner/Cargo.toml
@@ -1,12 +1,0 @@
-[package]
-name = "scanner"
-version = "0.1.0"
-authors = ["Nathaniel Cook <nvcook42@gmail.com>"]
-edition = "2018"
-
-
-[dependencies]
-
-[build-dependencies]
-cc = "1.0"
-bindgen = "0.49.0"
