--- internal/rust/semantic/Cargo.toml
@@ -1,10 +1,0 @@
-[package]
-name = "semantic"
-version = "0.1.0"
-authors = ["jlapacik <joshua.lapacik@pacbell.net>"]
-edition = "2018"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-maplit = "1.0.2"
