--- internal/rust/semantic/Makefile
@@ -1,11 +1,0 @@
-# Default target
-build:
-	$(CARGO) build
-
-clean:
-	$(CARGO) clean
-
-test:
-	$(CARGO) test
-
-.PHONY: build clean test
