+++ libflux/Cargo.toml
@@ -1,21 +1,28 @@
 [package]
-name = "flux-parser"
+name = "flux"
 version = "0.3.0"
 authors = ["Nathaniel Cook <nathaniel@influxdata.com>"]
 edition = "2018"
 
 [lib]
-crate-type = ["cdylib", "staticlib"]
+crate-type = ["rlib", "staticlib"]
 
 [dependencies]
+serde = "^1.0.59"
+serde_derive = "^1.0.59"
+serde_json = "1.0"
+serde-aux = "0.6.1"
 wasm-bindgen = { version = "0.2", features = ["serde-serialize"] }
 chrono = { version = "0.4", features = ["serde"] }
 regex = "1"
-scanner = {path = "../scanner" }
-ast = {path = "../ast" }
+maplit = "1.0.2"
 
 [dev-dependencies]
 pretty_assertions = "0.6.1"
+
+[build-dependencies]
+cc = "1.0"
+bindgen = "0.49.0"
 
 # https://rustwasm.github.io/docs/book/reference/code-size.html#optimizing-builds-for-code-size
 [profile.release]
