+++ libflux/Makefile
@@ -1,8 +1,11 @@
+export CARGO=cargo
 export CC=clang
 
 # Default target, build
 build:
 	wasm-pack build --dev --scope influxdata
+
+generate:
 
 clean:
 	$(CARGO) clean
