+++ libflux/build.sh
