+++ libflux/entry.sh
