+++ libflux/include/influxdata/flux.h
@@ -1,0 +1,12 @@
+#ifndef _INFLUXDATA_FLUX_H
+#define _INFLUXDATA_FLUX_H
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+#ifdef __cplusplus
+}
+#endif
+
+#endif
