+++ libflux/publish.sh
