+++ libflux/src/scanner/scanner.h
