+++ libflux/src/scanner/scanner.rl
