+++ libflux/src/scanner/tests.rs
