+++ libflux/src/scanner/unicode.rl
