+++ libflux/src/scanner/unicode.rl.COPYING
